Refactor: TMDB 외부 API 관련 유틸 함수 분리
